requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# API Comparison Script Dependencies
PyYAML>=6.0
requests>=2.31.0
